// packages/datadog-api-client-v1/models/SyntheticsDevice.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SyntheticsDeviceID } from './SyntheticsDeviceID';
/**
* Object describing the device used to perform the Synthetic test.
*/
export declare class SyntheticsDevice {
    /**
    * Screen height of the device.
    */
    'height': number;
    'id': SyntheticsDeviceID;
    /**
    * Whether or not the device is a mobile.
    */
    'isMobile'?: boolean;
    /**
    * The device name.
    */
    'name': string;
    /**
    * Screen width of the device.
    */
    'width': number;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SyntheticsDevice;
    static serialize(data: SyntheticsDevice): {
        [key: string]: any;
    };
    constructor();
}